Start splitting out tests for plugins.selection
